#1363 move XATransactionManagerSPILoader => transaction.xa.manager.
