+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/wrap/XATransactionDataSourceWrapperTest.java
@@ -22,7 +22,7 @@ import com.atomikos.jdbc.AtomikosDataSourceBean;
 import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.core.constant.transaction.ProxyPoolType;
 import io.shardingsphere.core.rule.DataSourceParameter;
-import io.shardingsphere.transaction.xa.XATransactionManagerSPILoader;
+import io.shardingsphere.transaction.xa.manager.XATransactionManagerSPILoader;
 import io.shardingsphere.transaction.xa.wrap.dialect.XADataSourceFactory;
 import io.shardingsphere.transaction.xa.manager.fixture.ReflectiveUtil;
 import io.shardingsphere.transaction.xa.wrap.dialect.XADatabaseType;
